+++ src/main/java/me/piitex/renjava/gui/overlay/ImageOverlay.java
@@ -2,5 +2,5 @@ package me.piitex.renjava.gui.overlay;
 
 import javafx.scene.image.Image;
 
-public record ImageOverlay(Image image, int x, int y) implements Overlay {
+public record ImageOverlay(Image image, double x, double y) implements Overlay {
 }
